refactor(db): split __integration__/setup.ts into setup/seed/cleanup
CR-local (round 2, major): setup.ts hit 307 lines, over the 200-line utility
cap (code-style.md §1) — and per the suppression-#5 clarification added earlier
in this PR, shared test-infra helpers are NOT covered by the .test.ts 500-line
exemption. Split into three focused, under-cap modules (verbatim moves, zero
logic change):
- setup.ts  (89L): env + clients + createTestOrg/createTestUser/getAuthenticatedClient
- seed.ts  (134L): seedReferenceData, seedQuestions
- cleanup.ts (85L): type ReferenceIds, cleanupTestData, cleanupReferenceData

All 8 integration suites updated to direct imports (no barrel file, per §4).
setup.test.ts split into seed.test.ts + cleanup.test.ts (co-located per source
file). 100 @repo/db tests pass; type-check clean.

Resolves the deferred split — closes #778.

Closes #778
