+++ packages/db/src/__integration__/cleanup.test.ts
@@ -1,10 +1,9 @@
 import { beforeEach, describe, expect, it, vi } from 'vitest'
 
 // ---------------------------------------------------------------------------
-// setup.ts calls requireEnv() at module scope — the env vars must be present
-// before the module is imported. vi.hoisted runs before any import, so we set
-// them here. The Supabase client factories (getAdminClient / getAnonClient)
-// are lazy functions; the only module-level side effect is requireEnv().
+// cleanup.ts does not call requireEnv() at module scope, but cleanup.test.ts
+// may co-load setup.ts indirectly (Vitest module graph). Set env vars here
+// via vi.hoisted to be safe — runs before any import.
 // ---------------------------------------------------------------------------
 
 vi.hoisted(() => {
@@ -23,7 +22,7 @@ vi.mock('@supabase/supabase-js', () => ({
 
 // vitest hoists the vi.mock calls above this import — the module gets the
 // mocked createClient and stubbed env vars before its top-level code runs.
-import { cleanupReferenceData, seedQuestions } from './setup'
+import { cleanupReferenceData } from './cleanup'
 
 // ---------------------------------------------------------------------------
 // buildChain — Proxy-based thenable that forwards every method call back to
@@ -45,93 +44,8 @@ function buildChain(returnValue: unknown): unknown {
   })
 }
 
-// Minimal admin mock: only `.from()` is needed — seedQuestions does not call
-// auth methods.
-const adminMock = { from: mockFrom } as unknown as Parameters<typeof seedQuestions>[0]['admin']
-
-const BASE_OPTS = {
-  admin: adminMock,
-  orgId: 'org-uuid-test',
-  createdBy: 'user-uuid-creator',
-  subjectId: 'subj-uuid-1',
-  topicId: 'topic-uuid-1',
-  count: 2,
-}
-
 beforeEach(() => {
   vi.resetAllMocks()
-})
-
-describe('seedQuestions', () => {
-  it('reuses the existing bank when one is found for the org', async () => {
-    const existingBankId = 'bank-uuid-existing'
-    const questionIds = [{ id: 'q-1' }, { id: 'q-2' }]
-
-    mockFrom
-      .mockReturnValueOnce(buildChain({ data: { id: existingBankId }, error: null })) // bank lookup → found
-      .mockReturnValueOnce(buildChain({ data: questionIds, error: null })) // questions insert
-
-    const result = await seedQuestions(BASE_OPTS)
-
-    expect(result.bankId).toBe(existingBankId)
-    expect(result.questionIds).toEqual(['q-1', 'q-2'])
-    // Only 2 .from() calls: lookup + questions insert (no bank insert)
-    expect(mockFrom).toHaveBeenCalledTimes(2)
-  })
-
-  it('creates a new bank when no bank exists for the org', async () => {
-    const newBankId = 'bank-uuid-new'
-    const questionIds = [{ id: 'q-3' }, { id: 'q-4' }]
-
-    mockFrom
-      .mockReturnValueOnce(buildChain({ data: null, error: null })) // bank lookup → not found
-      .mockReturnValueOnce(buildChain({ data: { id: newBankId }, error: null })) // bank insert
-      .mockReturnValueOnce(buildChain({ data: questionIds, error: null })) // questions insert
-
-    const result = await seedQuestions(BASE_OPTS)
-
-    expect(result.bankId).toBe(newBankId)
-    expect(result.questionIds).toEqual(['q-3', 'q-4'])
-    // 3 .from() calls: lookup + bank insert + questions insert
-    expect(mockFrom).toHaveBeenCalledTimes(3)
-  })
-
-  it('returns question IDs mapped from the inserted rows', async () => {
-    const questionIds = [{ id: 'q-a' }, { id: 'q-b' }, { id: 'q-c' }]
-
-    mockFrom
-      .mockReturnValueOnce(buildChain({ data: { id: 'bank-abc' }, error: null }))
-      .mockReturnValueOnce(buildChain({ data: questionIds, error: null }))
-
-    const result = await seedQuestions({ ...BASE_OPTS, count: 3 })
-
-    expect(result.questionIds).toHaveLength(3)
-    expect(result.questionIds).toEqual(['q-a', 'q-b', 'q-c'])
-  })
-
-  it('throws with "seedBank lookup:" prefix when the bank lookup query fails', async () => {
-    mockFrom.mockReturnValueOnce(
-      buildChain({ data: null, error: { message: 'connection refused' } }),
-    )
-
-    await expect(seedQuestions(BASE_OPTS)).rejects.toThrow(/seedBank lookup:/)
-  })
-
-  it('throws with "seedBank:" prefix when the bank insert fails', async () => {
-    mockFrom
-      .mockReturnValueOnce(buildChain({ data: null, error: null })) // lookup → not found
-      .mockReturnValueOnce(buildChain({ data: null, error: { message: 'unique violation' } })) // insert fails
-
-    await expect(seedQuestions(BASE_OPTS)).rejects.toThrow(/^seedBank:/)
-  })
-
-  it('throws with "seedQuestions:" prefix when the questions insert fails', async () => {
-    mockFrom
-      .mockReturnValueOnce(buildChain({ data: { id: 'bank-xyz' }, error: null })) // bank lookup OK
-      .mockReturnValueOnce(buildChain({ data: null, error: { message: 'FK violation' } })) // questions insert fails
-
-    await expect(seedQuestions(BASE_OPTS)).rejects.toThrow(/^seedQuestions:/)
-  })
 })
 
 // ---------------------------------------------------------------------------
