+++ packages/db/src/__integration__/setup.ts
@@ -87,220 +87,3 @@ export async function getAuthenticatedClient(opts: {
   if (error) throw new Error(`signIn: ${error.message}`)
   return client
 }
-
-/** Seed reference data: subject + topic + subtopic. Returns IDs. */
-export async function seedReferenceData(opts: {
-  admin: SupabaseClient
-  subjectCode: string
-  subjectName: string
-  topicCode: string
-  topicName: string
-  subtopicCode?: string
-  subtopicName?: string
-}): Promise<{ subjectId: string; topicId: string; subtopicId: string | null }> {
-  const { admin } = opts
-
-  const { data: subject, error: sErr } = await admin
-    .from('easa_subjects')
-    .upsert(
-      {
-        code: opts.subjectCode,
-        name: opts.subjectName,
-        short: opts.subjectCode,
-        sort_order: 1,
-      },
-      { onConflict: 'code' },
-    )
-    .select('id')
-    .single()
-  if (sErr) throw new Error(`seedSubject: ${sErr.message}`)
-
-  const { data: topic, error: tErr } = await admin
-    .from('easa_topics')
-    .upsert(
-      {
-        subject_id: subject.id,
-        code: opts.topicCode,
-        name: opts.topicName,
-        sort_order: 1,
-      },
-      { onConflict: 'subject_id,code' },
-    )
-    .select('id')
-    .single()
-  if (tErr) throw new Error(`seedTopic: ${tErr.message}`)
-
-  let subtopicId: string | null = null
-  if (opts.subtopicCode && opts.subtopicName) {
-    const { data: sub, error: stErr } = await admin
-      .from('easa_subtopics')
-      .upsert(
-        {
-          topic_id: topic.id,
-          code: opts.subtopicCode,
-          name: opts.subtopicName,
-          sort_order: 1,
-        },
-        { onConflict: 'topic_id,code' },
-      )
-      .select('id')
-      .single()
-    if (stErr) throw new Error(`seedSubtopic: ${stErr.message}`)
-    subtopicId = sub.id
-  }
-
-  return { subjectId: subject.id, topicId: topic.id, subtopicId }
-}
-
-/** Seed a question bank + questions. Returns bank ID and question IDs. */
-export async function seedQuestions(opts: {
-  admin: SupabaseClient
-  orgId: string
-  createdBy: string
-  subjectId: string
-  topicId: string
-  subtopicId?: string | null
-  count: number
-}): Promise<{ bankId: string; questionIds: string[] }> {
-  const { admin, orgId, createdBy, subjectId, topicId, subtopicId } = opts
-
-  // 1:1 org:bank invariant (mig 062) — reuse the existing bank if one exists,
-  // otherwise create one. Mirrors production lookup in insert-question.ts.
-  // Service-role bypasses RLS, so the soft-delete filter must be applied
-  // manually (production gets it from the authenticated-user RLS policy).
-  const { data: existingBank, error: lookupErr } = await admin
-    .from('question_banks')
-    .select('id')
-    .eq('organization_id', orgId)
-    .is('deleted_at', null)
-    .maybeSingle()
-  if (lookupErr) throw new Error(`seedBank lookup: ${lookupErr.message}`)
-
-  let bankId: string
-  if (existingBank) {
-    bankId = existingBank.id
-  } else {
-    const { data: bank, error: bErr } = await admin
-      .from('question_banks')
-      .insert({
-        organization_id: orgId,
-        name: `Test Bank ${Date.now()}`,
-        created_by: createdBy,
-      })
-      .select('id')
-      .single()
-    if (bErr) throw new Error(`seedBank: ${bErr.message}`)
-    bankId = bank.id
-  }
-
-  const questions = Array.from({ length: opts.count }, (_, i) => ({
-    organization_id: orgId,
-    bank_id: bankId,
-    subject_id: subjectId,
-    topic_id: topicId,
-    subtopic_id: subtopicId ?? null,
-    question_text: `Test question ${i + 1}?`,
-    options: [
-      { id: 'a', text: `Option A ${i}`, correct: false },
-      { id: 'b', text: `Option B ${i}`, correct: true },
-      { id: 'c', text: `Option C ${i}`, correct: false },
-      { id: 'd', text: `Option D ${i}`, correct: false },
-    ],
-    explanation_text: `Explanation for question ${i + 1}`,
-    difficulty: 'medium',
-    status: 'active',
-    created_by: createdBy,
-  }))
-
-  const { data, error: qErr } = await admin.from('questions').insert(questions).select('id')
-  if (qErr) throw new Error(`seedQuestions: ${qErr.message}`)
-
-  return {
-    bankId,
-    questionIds: (data as Array<{ id: string }>).map((q) => q.id),
-  }
-}
-
-/**
- * Clean up all test data created during a test run.
- * Deletes in reverse FK order using the service role client.
- */
-export async function cleanupTestData(opts: {
-  admin: SupabaseClient
-  orgId: string
-  userIds: string[]
-}) {
-  const { admin, orgId, userIds } = opts
-
-  // Delete in FK-safe order
-  await admin.from('audit_events').delete().eq('organization_id', orgId)
-  await admin.from('fsrs_cards').delete().in('student_id', userIds)
-  await admin.from('student_responses').delete().eq('organization_id', orgId)
-  const { data: sessionIds, error: sessionIdsErr } = await admin
-    .from('quiz_sessions')
-    .select('id')
-    .eq('organization_id', orgId)
-  if (sessionIdsErr) throw new Error(`cleanupTestData: quiz_sessions lookup failed: ${sessionIdsErr.message}`)
-  await admin
-    .from('quiz_session_answers')
-    .delete()
-    .in(
-      'session_id',
-      sessionIds?.map((s: { id: string }) => s.id) ?? [],
-    )
-  await admin.from('quiz_sessions').delete().eq('organization_id', orgId)
-  await admin.from('questions').delete().eq('organization_id', orgId)
-  await admin.from('question_banks').delete().eq('organization_id', orgId)
-  await admin.from('users').delete().in('id', userIds)
-  await admin.from('organizations').delete().eq('id', orgId)
-
-  // Delete auth users
-  for (const uid of userIds) {
-    await admin.auth.admin.deleteUser(uid)
-  }
-}
-
-type ReferenceIds = { subjectId: string; topicId: string; subtopicId: string | null }
-
-/**
- * Delete reference rows seeded by seedReferenceData(). easa_subjects/topics/subtopics
- * are GLOBAL tables (not org-scoped), so they're cleaned per-suite (not per-org).
- * Idempotent: dedups ids and .in() no-ops on already-removed rows. Null subtopicIds
- * are skipped (suites that seed no subtopic pass subtopicId: null).
- *
- * Accepts `undefined` ref entries and filters them out: a describe-scoped `let refs`
- * is `undefined` until beforeAll assigns it, and vitest still runs afterAll if beforeAll
- * throws. Filtering here (vs an `if (refs)` guard per call site) also covers the
- * multi-ref start-session case where an early seed succeeds but a later one throws
- * (e.g. [refs, undefined, undefined]) — a single call-site guard would miss that.
- */
-export async function cleanupReferenceData(opts: {
-  admin: SupabaseClient
-  refs: Array<ReferenceIds | undefined>
-}) {
-  const { admin } = opts
-  const refs = opts.refs.filter((r): r is ReferenceIds => r != null)
-  const subtopicIds = [...new Set(refs.map((r) => r.subtopicId).filter((v): v is string => v !== null))]
-  const topicIds = [...new Set(refs.map((r) => r.topicId))]
-  const subjectIds = [...new Set(refs.map((r) => r.subjectId))]
-
-  // FK-safe order: subtopics → topics → subjects (array order is load-bearing). All test
-  // rows that FK into easa_* (questions.{subject_id,topic_id,subtopic_id},
-  // quiz_sessions.subject_id, any exam_configs) are already removed by cleanupTestData,
-  // which every caller runs first. Sequential await preserves the delete order.
-  const targets: Array<{ table: string; label: string; ids: string[] }> = [
-    { table: 'easa_subtopics', label: 'subtopics', ids: subtopicIds },
-    { table: 'easa_topics', label: 'topics', ids: topicIds },
-    { table: 'easa_subjects', label: 'subjects', ids: subjectIds },
-  ]
-  for (const { table, label, ids } of targets) {
-    if (ids.length === 0) continue
-    // Chain .select('id') per code-style.md §5: zero rows is a valid steady state, so only
-    // log when rows actually changed — surfaces a filter/ID regression that would no-op.
-    const { data: deleted, error } = await admin.from(table).delete().in('id', ids).select('id')
-    if (error) throw new Error(`cleanupReferenceData ${label}: ${error.message}`)
-    if ((deleted?.length ?? 0) > 0) {
-      console.log(`[cleanupReferenceData] removed ${deleted?.length} ${table}`)
-    }
-  }
-}
